+++ Dyalect/Runtime/Types/DyString.cs
@@ -61,7 +61,7 @@ namespace Dyalect.Runtime.Types
     {
         public static readonly DyStringTypeInfo Instance = new DyStringTypeInfo();
 
-        private DyStringTypeInfo() : base(StandardType.Bool, false)
+        private DyStringTypeInfo() : base(StandardType.String, false)
         {
 
         }
@@ -116,10 +116,175 @@ namespace Dyalect.Runtime.Types
 
         protected override DyString ToStringOp(DyObject arg, ExecutionContext ctx) => StringUtil.Escape(arg.GetString());
 
+        private DyObject IndexOf(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var a = args.TakeOne(DyNil.Instance);
+            var str = self.GetString();
+
+            if (a.TypeId == StandardType.String)
+                return DyInteger.Get(str.IndexOf(a.GetString()));
+            else if (a.TypeId == StandardType.Char)
+                return DyInteger.Get(str.IndexOf(a.GetChar()));
+            else
+                return Err.InvalidType(StandardType.CharName, a.TypeName(ctx)).Set(ctx);
+        }
+
+        private DyObject LastIndexOf(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var a = args.TakeOne(DyNil.Instance);
+            var str = self.GetString();
+
+            if (a.TypeId == StandardType.String)
+                return DyInteger.Get(str.LastIndexOf(a.GetString()));
+            else if (a.TypeId == StandardType.Char)
+                return DyInteger.Get(str.LastIndexOf(a.GetChar()));
+            else
+                return Err.InvalidType(StandardType.CharName, a.TypeName(ctx)).Set(ctx);
+        }
+
+        private DyObject Split(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var allChars = true;
+
+            for (var i = 0; i < args.Length; i++)
+                if (args[i].TypeId != StandardType.Char)
+                {
+                    allChars = false;
+                    break;
+                }
+
+            return allChars ? SplitByChars(ctx, self, args) : SplitByStrings(ctx, self, args);
+        }
+
+        private DyObject SplitByStrings(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var xs = new string[args.Length];
+
+            for (var i = 0; i < args.Length; i++)
+            {
+                if (args[i].TypeId != StandardType.String)
+                    return Err.InvalidType(StandardType.StringName, args[i].TypeName(ctx)).Set(ctx);
+
+                xs[i] = args[i].GetString();
+            }
+
+            var arr = self.GetString().Split(xs, StringSplitOptions.RemoveEmptyEntries);
+            var list = new List<DyObject>(arr.Length);
+
+            for (var i = 0; i < arr.Length; i++)
+                list.Add(new DyString(arr[i]));
+
+            return new DyArray(list);
+        }
+
+        private DyObject SplitByChars(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var xs = new char[args.Length];
+
+            for (var i = 0; i < args.Length; i++)
+                xs[i] = args[i].GetChar();
+
+            var arr = self.GetString().Split(xs, StringSplitOptions.RemoveEmptyEntries);
+            var list = new List<DyObject>(arr.Length);
+
+            for (var i = 0; i < arr.Length; i++)
+                list.Add(new DyString(arr[i]));
+
+            return new DyArray(list);
+        }
+
+        private DyObject Upper(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            return new DyString(self.GetString().ToUpper());
+        }
+
+        private DyObject Lower(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            return new DyString(self.GetString().ToLower());
+        }
+
+        private DyObject StartsWith(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var a = args.TakeOne(DyNil.Instance);
+
+            if (a.TypeId == StandardType.String)
+                return self.GetString().StartsWith(a.GetString()) ? DyBool.True : DyBool.False;
+
+            if (a.TypeId == StandardType.Char)
+                return self.GetString().StartsWith(a.GetChar()) ? DyBool.True : DyBool.False;
+
+            return Err.InvalidType(StandardType.StringName, a.TypeName(ctx)).Set(ctx);
+        }
+
+        private DyObject EndsWith(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var a = args.TakeOne(DyNil.Instance);
+
+            if (a.TypeId == StandardType.String)
+                return self.GetString().EndsWith(a.GetString()) ? DyBool.True : DyBool.False;
+
+            if (a.TypeId == StandardType.Char)
+                return self.GetString().EndsWith(a.GetChar()) ? DyBool.True : DyBool.False;
+
+            return Err.InvalidType(StandardType.StringName, a.TypeName(ctx)).Set(ctx);
+        }
+
+        private DyObject Substring(ExecutionContext ctx, DyObject self, DyObject[] args)
+        {
+            var from = args.TakeOne(DyNil.Instance);
+            var to = args.TakeAt(1, null);
+
+            if (from.TypeId != StandardType.Integer)
+                return Err.InvalidType(StandardType.IntegerName, from.TypeName(ctx)).Set(ctx);
+
+            if (to != null && to.TypeId != StandardType.Integer)
+                return Err.InvalidType(StandardType.IntegerName, to.TypeName(ctx)).Set(ctx);
+
+            var str = self.GetString();
+            var i = (int)from.GetInteger();
+
+            if (i < 0 || i >= str.Length)
+                return Err.IndexOutOfRange(self.TypeName(ctx), i).Set(ctx);
+
+            if (to == null)
+                return new DyString(str.Substring(i));
+
+            var j = (int)to.GetInteger();
+
+            if (j < 0 || j + i > str.Length)
+                return Err.IndexOutOfRange(self.TypeName(ctx), j).Set(ctx);
+
+            return new DyString(self.GetString().Substring(i, j));
+        }
+
         protected override DyFunction GetMember(string name, ExecutionContext ctx)
         {
             if (name == Builtins.Len)
                 return DyForeignFunction.Create(name, LenAdapter);
+
+            if (name == "indexOf")
+                return DyForeignFunction.Create(name, IndexOf);
+
+            if (name == "lastIndexOf")
+                return DyForeignFunction.Create(name, LastIndexOf);
+
+            if (name == "split")
+                return DyForeignFunction.Create(name, Split);
+
+            if (name == "upper")
+                return DyForeignFunction.Create(name, Upper);
+
+            if (name == "lower")
+                return DyForeignFunction.Create(name, Lower);
+
+            if (name == "startsWith")
+                return DyForeignFunction.Create(name, StartsWith);
+
+            if (name == "endsWith")
+                return DyForeignFunction.Create(name, EndsWith);
+
+            if (name == "sub")
+                return DyForeignFunction.Create(name, Substring);
 
             return null;
         }